+++ src/main/java/pl/wojciechkarpiel/jhou/ast/Equality.java
@@ -1,5 +1,6 @@
 package pl.wojciechkarpiel.jhou.ast;
 
+import pl.wojciechkarpiel.jhou.alpha.AlphaEqual;
 import pl.wojciechkarpiel.jhou.normalizer.Normalizer;
 
 public class Equality {
@@ -7,12 +8,10 @@ public class Equality {
     }
 
     public static boolean alphaEqual(Term a, Term b) {
-        return a.equals(b);
+        return AlphaEqual.isAlphaEqual(a, b);
     }
 
     public static boolean alphaBetaEtaEqual(Term a, Term b) {
-        // instead of eta-contracting a beta-normal form, we could beta-eta normalize
-        // former option seems much more straightforward, both should be equivalent
         return alphaEqual(
                 Normalizer.etaContract(Normalizer.betaNormalize(a)),
                 Normalizer.etaContract(Normalizer.betaNormalize(b))