+++ src/test/java/pl/wojciechkarpiel/jhou/alpha/AlphaEqualTest.java
@@ -5,9 +5,7 @@ import pl.wojciechkarpiel.jhou.ast.Abstraction;
 import pl.wojciechkarpiel.jhou.ast.Variable;
 import pl.wojciechkarpiel.jhou.ast.type.BaseType;
 import pl.wojciechkarpiel.jhou.ast.type.Type;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import pl.wojciechkarpiel.jhou.testUtil.TestUtil;
 
 class AlphaEqualTest {
 
@@ -16,9 +14,9 @@ class AlphaEqualTest {
         Type t = BaseType.freshBaseType();
         Variable x = Variable.freshVariable(t, "x");
         Variable y = Variable.freshVariable(t, "y");
-        assertEquals(new Abstraction(x, x), new Abstraction(y, y));
-        assertNotEquals(new Abstraction(x, y), new Abstraction(x, x));
-        assertNotEquals(new Abstraction(y, x), new Abstraction(x, x));
-        assertNotEquals(new Abstraction(y, x), new Abstraction(x, y));
+        TestUtil.assertAlphaEqual(new Abstraction(x, x), new Abstraction(y, y));
+        TestUtil.assertNotAlphaEqual(new Abstraction(x, y), new Abstraction(x, x));
+        TestUtil.assertNotAlphaEqual(new Abstraction(y, x), new Abstraction(x, x));
+        TestUtil.assertNotAlphaEqual(new Abstraction(y, x), new Abstraction(x, y));
     }
 }